pyproject.toml
[tool.poetry]
name = "demo-backend"
version = "0.1.0"
description = ""
authors = ["Tom Whitwell <tom@whi.tw>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask-RESTful = "^0.3.9"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
